move probing of existing issues into `gh`
