+++ mdbook-goals/src/gh.rs
@@ -3,5 +3,5 @@
 //! We do most everything through the `gh` command-line tool.
 
 pub mod issue_id;
-
+pub mod issues;
 pub mod labels;
