+++ mdbook-goals/src/rfc.rs
@@ -1,5 +1,5 @@
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::BTreeSet,
     fmt::Display,
     path::{Path, PathBuf},
     process::Command,
@@ -8,11 +8,11 @@ use std::{
 
 use anyhow::Context;
 use regex::Regex;
-use serde::{Deserialize, Serialize};
 
 use crate::{
     gh::{
         issue_id::IssueId,
+        issues::{list_issue_titles_in_milestone, ExistingGithubIssue},
         labels::{list_labels, GhLabel},
     },
     goal::{self, GoalDocument, ParsedOwners, PlanItem, Status},
@@ -194,86 +194,6 @@ enum GithubAction<'doc> {
     },
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct ExistingGithubIssue {
-    number: u64,
-    /// Just github username, no `@`
-    assignees: BTreeSet<String>,
-    comments: Vec<ExistingGithubComment>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct ExistingGithubComment {
-    /// Just github username, no `@`
-    author: String,
-    body: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct ExistingGithubIssueJson {
-    title: String,
-    number: u64,
-    assignees: Vec<ExistingGithubAssigneeJson>,
-    comments: Vec<ExistingGithubCommentJson>,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct ExistingGithubAssigneeJson {
-    login: String,
-    name: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct ExistingGithubCommentJson {
-    body: String,
-    author: ExistingGithubAuthorJson,
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-struct ExistingGithubAuthorJson {
-    login: String,
-}
-
-fn list_issue_titles_in_milestone(
-    repository: &str,
-    timeframe: &str,
-) -> anyhow::Result<BTreeMap<String, ExistingGithubIssue>> {
-    let output = Command::new("gh")
-        .arg("-R")
-        .arg(repository)
-        .arg("issue")
-        .arg("list")
-        .arg("-m")
-        .arg(timeframe)
-        .arg("-s")
-        .arg("all")
-        .arg("--json")
-        .arg("title,assignees,number,comments")
-        .output()?;
-
-    let existing_issues: Vec<ExistingGithubIssueJson> = serde_json::from_slice(&output.stdout)?;
-
-    Ok(existing_issues
-        .into_iter()
-        .map(|e_i| {
-            (
-                e_i.title,
-                ExistingGithubIssue {
-                    number: e_i.number,
-                    assignees: e_i.assignees.into_iter().map(|a| a.login).collect(),
-                    comments: e_i
-                        .comments
-                        .into_iter()
-                        .map(|c| ExistingGithubComment {
-                            author: format!("@{}", c.author.login),
-                            body: c.body,
-                        })
-                        .collect(),
-                },
-            )
-        })
-        .collect())
-}
 /// Initializes the required `T-<team>` labels on the repository.
 /// Warns if the labels are found with wrong color.
 fn initialize_labels(
@@ -373,14 +293,6 @@ fn initialize_issues<'doc>(
     }
 
     Ok(actions)
-}
-
-impl ExistingGithubIssue {
-    /// We use the presence of a "lock comment" as a signal that we successfully locked the issue.
-    /// The github CLI doesn't let you query that directly.
-    fn was_locked(&self) -> bool {
-        self.comments.iter().any(|c| c.body.trim() == LOCK_TEXT)
-    }
 }
 
 fn issue<'doc>(timeframe: &str, document: &'doc GoalDocument) -> anyhow::Result<GithubIssue<'doc>> {
@@ -691,4 +603,12 @@ fn lock_issue(repository: &str, number: u64) -> anyhow::Result<()> {
 /// Returns a comma-separated list of the strings in `s` (no spaces).
 fn comma(s: &BTreeSet<String>) -> String {
     s.iter().map(|s| &s[..]).collect::<Vec<_>>().join(",")
+}
+
+impl ExistingGithubIssue {
+    /// We use the presence of a "lock comment" as a signal that we successfully locked the issue.
+    /// The github CLI doesn't let you query that directly.
+    fn was_locked(&self) -> bool {
+        self.comments.iter().any(|c| c.body.trim() == LOCK_TEXT)
+    }
 }
